Change error message when reading tokens
In my opinion, the previous message is a bit confusing, we know that only "alse"
can follow "f", so we expect that to follow, so let's say we expect something
instead of saying that something is incomplete. Also, "false" is not a literal
in this context, for the user, it's a value.